refactor: corrected remarks after PR
