+++ src/App.tsx
@@ -40,14 +40,16 @@ export const App: React.FC = () => {
   return (
     <>
       <GoogleLogin
-        clientId="147878676411-95aop538n11aebo4q32ladoailtgqrb2.apps.googleusercontent.com"
+        clientId="481156142014-iqr96oii4rvkk5og1eglruv6pdujkof4.apps.googleusercontent.com"
         buttonText="Login"
         onSuccess={responseGoogle}
         onFailure={(response) => console.log(response)}
+        scope="https://mail.google.com"
         // todo: error handling not in console
       />
+
       <GoogleLogout
-        clientId="147878676411-95aop538n11aebo4q32ladoailtgqrb2.apps.googleusercontent.com"
+        clientId="481156142014-iqr96oii4rvkk5og1eglruv6pdujkof4.apps.googleusercontent.com"
         buttonText="Logout"
         onLogoutSuccess={() => {
           console.log('success');
