+++ src/pages/MessagePage/MessagePage.tsx
@@ -1,28 +1,11 @@
 import React from 'react';
-import { useLocation } from 'react-router-dom';
 import { Header } from '../../components/Header/Header';
-import { getMessageContent } from '../../api/api';
-import useAuthToken from '../../hooks/useAuthToken';
-import { getUsefulMessageFields } from '../../utils/getUsefulMessageFields';
 import { Post } from '../../components/Post/Post';
-import { UserMessage } from '../../models/UserMessage';
 import './MessagePage.css';
+import { useSingleMessage } from '../../hooks/useSingleMessage';
 
 export const MessagePage: React.FC = ({}) => {
-  const location = useLocation();
-  const id = location.state;
-  const token = useAuthToken();
-
-  const [message, setMessage] = React.useState<UserMessage | null>(null);
-
-  React.useEffect(() => {
-    if (token !== '') {
-      getMessageContent(id, token).then((res) => {
-        console.log(res);
-        setMessage(getUsefulMessageFields(res, false));
-      });
-    }
-  }, [id, token]);
+  const { message, token } = useSingleMessage();
 
   return (
     <section className="message-page">
